requirements.txt
streamlit==1.43.2
GitPython~=3.1.46
gitdb~=4.0.12
numpy~=2.4.4
pillow~=11.3.0
pip~=25.2
typing_extensions~=4.15.0
attrs~=26.1.0
protobuf~=5.29.6
cryptography~=46.0.6
pytz~=2026.1.post1
toml~=0.10.2
smmap~=5.0.3
charset-normalizer~=3.4.7
cffi~=2.0.0
altair~=5.5.0
jsonschema~=4.26.0
narwhals~=2.18.1
pandas~=2.3.3
pyarrow~=23.0.1
Jinja2~=3.1.6
packaging~=24.2
referencing~=0.37.0
MarkupSafe~=3.0.3
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
certifi~=2026.2.25
tzdata~=2026.1
tornado~=6.5.5
colorama~=0.4.6
urllib3~=2.6.3
six~=1.17.0
idna~=3.11
tenacity~=9.1.4
watchdog~=6.0.0
click~=8.3.2
cachetools~=5.5.2
blinker~=1.9.0
jsonschema-specifications~=2025.9.1
PyYAML~=6.0.3